requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
Django==5.2.1
django-cloudinary-storage==0.3.0
django-environ==0.12.0
django-js-asset==3.1.2
django-markdownx==4.0.9
gunicorn==23.0.0
idna==3.10
Markdown==3.8
packaging==25.0
pillow==11.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
